data/data_fetcher.py
import asyncio
import logging
from typing import Dict, Any
from datetime import datetime

logger = logging.getLogger(__name__)

class DataFetcher:
    """
    Handles fetching and processing of market data from various sources.
    """
    
    def __init__(self, config: Dict[str, Any]):
        """
        Initialize the DataFetcher with configuration settings.
        
        Args:
            config: Dictionary containing configuration parameters
        """
        self.config = config
        self.logger = logging.getLogger(__name__)
        self.logger.info("Initializing DataFetcher")

    async def process_data_async(self, raw_data: Dict[str, Any]) -> Dict[str, Any]:
        """
        Process raw market data asynchronously.
        
        Args:
            raw_data: Dictionary containing raw market data
            
        Returns:
            Dictionary containing processed market data
        """
        try:
            self.logger.debug("Processing raw market data")
            
            # Extract relevant fields
            processed_data = {
                "timestamp": datetime.now().isoformat(),
                "symbol": raw_data.get("symbol"),
                "price": raw_data.get("price"),
                "volume": raw_data.get("volume"),
                "bid": raw_data.get("bid"),
                "ask": raw_data.get("ask"),
                "high": raw_data.get("high"),
                "low": raw_data.get("low"),
                "open": raw_data.get("open"),
                "close": raw_data.get("close")
            }

            # Calculate derived metrics
            if all(key in processed_data for key in ["bid", "ask"]):
                processed_data["spread"] = processed_data["ask"] - processed_data["bid"]

            if all(key in processed_data for key in ["high", "low"]):
                processed_data["range"] = processed_data["high"] - processed_data["low"]

            self.logger.debug("Data processing complete")
            return processed_data

        except Exception as e:
            self.logger.error(f"Error processing market data: {str(e)}")
            raise

    async def fetch_historical_data(self, symbol: str, timeframe: str, limit: int = 100) -> Dict[str, Any]:
        """
        Fetch historical market data for backtesting and analysis.
        
        Args:
            symbol: Trading symbol
            timeframe: Time interval for the data
            limit: Number of historical data points to fetch
            
        Returns:
            Dictionary containing historical market data
        """
        try:
            self.logger.info(f"Fetching historical data for {symbol}")
            # Implementation for historical data fetching would go here
            # This would typically involve API calls to the data provider
            
            # Placeholder return
            return {
                "symbol": symbol,
                "timeframe": timeframe,
                "data": []
            }
            
        except Exception as e:
            self.logger.error(f"Error fetching historical data: {str(e)}")
            raise

    async def validate_data(self, data: Dict[str, Any]) -> bool:
        """
        Validate market data for completeness and accuracy.
        
        Args:
            data: Dictionary containing market data to validate
            
        Returns:
            Boolean indicating if data is valid
        """
        required_fields = ["symbol", "price", "timestamp"]
        
        try:
            # Check for required fields
            if not all(field in data for field in required_fields):
                self.logger.warning("Missing required fields in market data")
                return False
                
            # Validate data types and ranges
            if not isinstance(data["price"], (int, float)) or data["price"] <= 0:
                self.logger.warning("Invalid price in market data")
                return False
                
            return True
            
        except Exception as e:
            self.logger.error(f"Error validating market data: {str(e)}")
            return False
